Refactoring and improved Bugfix
Since the ENTRY_MODIFY event triggers twice (one for the content change, and one
for the timestamp change) a simple sleep was added to prevent that. However this
soultion is very poor since we might not get the same behaviour on all machines.
Therefore added a boolan flag to determine multiple ENTRY_MODIFY events.